New reader for SINTEF Ocean data exchange format based on matlab .mat files. Supports all versions <=7.2 and >=7.3.
