+++ qats/readers/__init__.py
@@ -4,7 +4,7 @@ Sub-package with readers for various file formats.
 """
 from . import csv
 from . import direct_access
-from . import matlab
+from . import sintef_mat
 from . import other
 from . import sima
 from . import sima_h5
